mcp_server/requirements.txt
mcp
psycopg[binary]
python-dotenv
openai
streamlit
requests
pandas
